Updated/Data-Science-Capstone-Website-updated/requirements.txt
fastapi==0.104.1
fastjsonschema==2.18.0
fastparquet==2023.8.0
filelock==3.12.4
flatbuffers==23.5.26
fonttools==4.42.1
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.9.0
gitdb==4.0.11
GitPython==3.1.40
gmpy2 
greenlet==3.0.1
grpcio==1.59.3
grpcio-status==1.59.3
h11==0.14.0
httptools==0.6.1
huggingface-hub==0.17.1
humanfriendly==10.0
idna 
ipykernel
ipython 
ipywidgets==8.1.1
isoduration==20.11.0
jedi
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client
jupyter_core 
jupyter_server==2.7.3
jupyter_server_terminals==0.4.4
jupyterlab==4.0.6
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
jupyterlab_server==2.25.0
kiwisolver==1.4.5
kubernetes==28.1.0
libclang==16.0.6
markdown-it-py==3.0.0
markdown2==2.4.12
MarkupSafe 
mdurl==0.1.2
mistune==3.0.2
mkl-fft 
mkl-random 
mkl-service==2.4.0
mmh3==4.0.1
monotonic==1.6
mpmath 
multidict==6.0.4
multiprocess 
mypy-extensions==1.0.0
mysql-connector-python==8.2.0
nbclient==0.8.0
nbconvert==7.8.0
nbformat==5.9.2
nest-asyncio
networkx
notebook==7.0.3
notebook_shim==0.2.3
numpy==1.24.3
oauthlib==3.2.2
onnxruntime==1.16.3
openpyxl==3.1.2
overrides==7.4.0
packaging 
pandas
pandocfilters==1.5.0
parso 
Pillow==10.0.1
platformdirs 
posthog==3.0.2
prometheus-client==0.17.1
prompt-toolkit 
proto-plus==1.22.3
protobuf==4.21.12
psutil 
pulsar-client==3.3.0
pure-eval 
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
pydeck==0.8.1b0
Pygments 
PyMySQL==1.1.0
pyparsing 
PyPika==0.48.9
pyreadline3==3.4.1
PySocks 
python-dateutil 
python-dotenv==1.0.0
python-json-logger==2.0.7
pytz==2023.3.post1
PyYAML 
pyzmq 
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.6.0
rpds-py==0.10.3
rsa==4.9
safetensors==0.3.3
scikit-learn==1.3.0
scipy 
Send2Trash==1.8.2
six 
smmap==5.0.1
sniffio==1.3.0
snuggs 
soupsieve==2.5
SQLAlchemy==2.0.23
stack-data 
starlette==0.27.0
streamlit==1.32.2
sympy 
termcolor==2.3.0
terminado==0.17.1
threadpoolctl==3.2.0
tinycss2==1.2.1
toml==0.10.2
toolz==0.12.0
tornado
traitlets 
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
uri-template==1.3.0
urllib3==1.26.16
uvicorn==0.24.0.post1
validators==0.22.0
watchdog==3.0.0
watchfiles==0.21.0
wcwidth
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.3
websockets==12.0
widgetsnbextension==4.0.9
win-inet-pton 
wrapt==1.15.0
yarl==1.9.3
